convert neurosift.js to a ts file
